+++ apps/prairielearn/src/pages/instructorAssessmentInstances/instructorAssessmentInstances.html.ts
@@ -33,7 +33,7 @@ export function InstructorAssessmentInstances({ resLocals }: { resLocals: Record
           rel="stylesheet"
         />
 
-        ${compiledScriptTag('instructorAssessmentInstancesClient.ts')}
+        ${compiledScriptTag('instructorAssessmentInstancesClient.tsx')}
       </head>
       <body>
         ${Navbar({ resLocals })}
@@ -169,6 +169,7 @@ export function InstructorAssessmentInstances({ resLocals }: { resLocals: Record
               data-csrf-token="${resLocals.__csrf_token}"
               data-has-course-instance-permission-edit="${resLocals.authz_data
                 .has_course_instance_permission_edit}"
+              data-timezone="${resLocals.course_instance.display_timezone}"
             ></table>
 
             <div class="spinning-wheel card-body spinner-border">
@@ -402,8 +403,7 @@ function GradeAllAssessmentInstancesModal({
     title: 'Grade all assessment instances',
     body: html`
       Are you sure you want to grade pending submissions for all assessment instances for
-      <strong> ${assessmentSetName} ${assessmentNumber} </strong>
-      ? This cannot be undone.
+      <strong>${assessmentSetName} ${assessmentNumber}</strong>? This cannot be undone.
     `,
     footer: html`
       <input type="hidden" name="__action" value="grade_all" />
@@ -428,8 +428,7 @@ function CloseAllAssessmentInstancesModal({
     title: 'Grade and Close all assessment instances',
     body: html`
       Are you sure you want to grade and close all assessment instances for
-      <strong> ${assessmentSetName} ${assessmentNumber} </strong>
-      ? This cannot be undone.
+      <strong>${assessmentSetName} ${assessmentNumber}</strong>? This cannot be undone.
     `,
     footer: html`
       <input type="hidden" name="__action" value="close_all" />