+++ packages/run/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/run
 
+## 1.0.3
+
+### Patch Changes
+
+- 9d7d790: Upgrade all JavaScript dependencies
+
+## 1.0.2
+
+### Patch Changes
+
+- 315d931: Upgrade all JavaScript dependencies
+
 ## 1.0.1
 
 ### Patch Changes