+++ packages/sentry/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/sentry
 
+## 2.0.11
+
+### Patch Changes
+
+- 9d7d790: Upgrade all JavaScript dependencies
+
+## 2.0.10
+
+### Patch Changes
+
+- 85300c4: Upgrade all JavaScript dependencies
+
+## 2.0.9
+
+### Patch Changes
+
+- 315d931: Upgrade all JavaScript dependencies
+
 ## 2.0.8
 
 ### Patch Changes